ingestion/requirements.txt
kafka-python
yfinance
